Bug2 correction
Bug2 occured because I had not modified the part of the code that sets
the pcr_packet_count and pcr_packet period. In the original application,
since there was only one service, it didn't matter which stream was used
to test if there already was a PID with PCR. When another service was
added, the application looked to the last stream on the stream list and
set the pcr_* fields to the service which contained that stream, and
just did nothing to the other service, so the pcr_period was '0'  and
a PCR stamp was added in every TS packet that was supposed to carry the
PCR, and not only in some packets. The solution was acheived by creating
a loop over all services and assigning the control variables based on
the service list inside the 'ts' structure and not in the pointer of
the 'ts_st' structure as formerly.